Some new post meta fields: draft and date
Not all of the ones wanted in #11, but two for three.

There are now two new important meta fields for posts (neither is required),
draft and date. Draft specifies if something is a draft or not (as in, put
`draft: true` if something is a draft. You can omit it if you want). Date
specifies the post's timestamp, overwriting the one in the filename.

There are some things I still want to work on, specifically with post filenames,
such as respecting directories (if they're, you know, a date order:
(<year>/<month>/<day>/my-cool-post.md), and even removing the need for a date
anywhere in the file's name if it's set in the meta. But, that's stuff for
later.

So, that's what's new...