Did someone say "Generate drafts?"
In reference to #12, drafts are now generated. There are some more things that
probably need to be setup, and added to the main configuration file, but, right
now, drafts are generated to `/drafts/<title>/`. If your sever supports
directory browsing, going to that url will display all of your current drafts
(located somewhere in the `_posts` directory, provided they have the YYYY-MM-DD
at the beginning of their filename (otherwise fjord breaks. This is something I
want to fix, and it'll be part of the big `_posts` directory overhaul (which
should also include custom post stuff, which'll be nice (see #9))).

Look, parentheticals in parentheticals.